breaking up constraints into a classes file and a dsl file
