+++ lib/chef/sugar/constraints.rb
@@ -17,34 +17,6 @@
 class Chef
   module Sugar
     module Constraints
-      extend self
-
-      #
-      # Shortcut method for creating a new {Version} object.
-      #
-      # @param [String] version
-      #   the version (as a string) to create
-      #
-      # @return [Chef::Sugar::Constraints::Version]
-      #   the new version object
-      #
-      def version(version)
-        Chef::Sugar::Constraints::Version.new(version)
-      end
-
-      #
-      # Shortcut method for creating a new {Constraint} object.
-      #
-      # @param [String, Array<String>] constraints
-      #   the list of constraints to use
-      #
-      # @return [Chef::Sugar::Constraints::Constraint]
-      #   the new constraint object
-      #
-      def constraint(*constraints)
-        Chef::Sugar::Constraints::Constraint.new(*constraints)
-      end
-
       #
       # This class is a wrapper around a version requirement that adds a nice
       # DSL for comparing constraints:
@@ -129,32 +101,6 @@ class Chef
         def satisfies?(*constraints)
           Gem::Requirement.new(*constraints).satisfied_by?(@version)
         end
-      end
-    end
-
-    module DSL
-      # @see Chef::Sugar::Constraints#version
-      def version(version)
-        Chef::Sugar::Constraints::Version.new(version)
-      end
-
-      # @see Chef::Sugar::Constraints#constraint
-      def constraint(*constraints)
-        Chef::Sugar::Constraints.constraint(*constraints)
-      end
-
-      #
-      # This wrapper/convenience method is only available in the recipe DSL. It
-      # creates a new version object from the {Chef::VERSION}.
-      #
-      # @example Check if Chef 11+
-      #   chef_version.satisfies?('>= 11.0.0')
-      #
-      # @return [Chef::Sugar::Constraints::Version]
-      #   a version object, wrapping the current {Chef::VERSION}
-      #
-      def chef_version
-        version(Chef::VERSION)
       end
     end
   end
